Moved inventory related classes to Inventory folder
