+++ Source/UE4CPP/Inventory/Interfaces/InventoryItemBPI.cpp
@@ -2,5 +2,5 @@
 
 void IInventoryItemBPI::GetInventoryItem_Implementation(FInventoryItem& InventoryItem)
 {
-    InventoryItem = FInventoryItem(0, InventoryItemType::Default, FIntVector(1,1,1), "Name", "");
+    InventoryItem = FInventoryItem(0, EInventoryItemType::Default, FIntVector(1,1,1), "Name", "");
 }
