+++ Source/UE4CPP/Inventory/InventoryActorComponent.h
