+++ Source/UE4CPP/Inventory/Structs/Inventory.h
