+++ Source/UE4CPP/Inventory/Structs/InventoryItem.cpp
